修改了dubbo服务的端口，使用了yml配置。现在服务运行通畅

+++ b2b-dubbo-cart/src/main/java/com/b2b/dubbo/cart/service/CartService.java
@@ -6,10 +6,10 @@ import com.b2b.mall.db.model.Item;
 import java.util.List;
 
 public interface CartService {
-    Result addCart(Long userId, Long itemId, int num);
+    Result addCart(Long userId, Integer itemId, int num);
     Result mergeCart(Long userId, List<Item> cookieItemList);
     List<Item> getCartList(Long userId);
-    Result updateCartNum(Long userId, Long itemId, int num);
-    Result deleteCartItem(Long userId, Long itemId);
+    Result updateCartNum(Long userId, Integer itemId, int num);
+    Result deleteCartItem(Long userId, Integer itemId);
     Result clearCartList(Long userId);
 }
